% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corona_calculations.R
\name{CORONA}
\alias{CORONA}
\alias{create_hover_info_corona}
\title{Stock Calculations}
\usage{
CORONA(country, datestart, dateend)

create_hover_info_corona(hovercorona, coronadata, selectedmeasure)
}
\description{
Stock Calculations
}
