% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_load_stockdata.R
\name{1COV}
\alias{1COV}
\alias{ADS}
\alias{ALV}
\alias{BAS}
\alias{BAYN}
\alias{BEI}
\alias{BMW}
\alias{CON}
\alias{DAI}
\alias{DB1}
\alias{DBK}
\alias{DHER}
\alias{DPW}
\alias{DTE}
\alias{DWNI}
\alias{EOAN}
\alias{FME}
\alias{FRE}
\alias{GDAXI}
\alias{HEI}
\alias{HEN3}
\alias{IFX}
\alias{COMPONENTS_DE}
\alias{LIN}
\alias{MRK}
\alias{MTX}
\alias{MUV2}
\alias{RWE}
\alias{SAP}
\alias{SIE}
\alias{VNA}
\alias{VOW3}
\title{Functions to load stock data}
\usage{
`1COV`()

ADS()

ALV()

BAS()

BAYN()

BEI()

BMW()

CON()

DAI()

DB1()

DBK()

DHER()

DPW()

DTE()

DWNI()

EOAN()

FME()

FRE()

GDAXI()

HEI()

HEN3()

IFX()

COMPONENTS_DE()

LIN()

MRK()

MTX()

MUV2()

RWE()

SAP()

SIE()

VNA()

VOW3()
}
\description{
Functions to load stock data
}
