% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{selectize_Stocks}
\alias{selectize_Stocks}
\alias{sliderinput_dates}
\alias{plot_stocks_DE}
\alias{hover_info_DE}
\alias{selectize_corona}
\alias{sliderinput_dates_corona}
\alias{plot_corona}
\alias{hover_info_corona}
\title{UI Elements}
\usage{
selectize_Stocks(components)

sliderinput_dates()

plot_stocks_DE()

hover_info_DE()

selectize_corona()

sliderinput_dates_corona()

plot_corona()

hover_info_corona()
}
\description{
UI Elements
}
