% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{selectize_Stocks}
\alias{selectize_Stocks}
\alias{sliderinput_dates}
\alias{plot_stocks_DE}
\alias{hover_info_DE}
\alias{plot_stocks_US}
\alias{hover_info_US}
\title{UI Elements}
\usage{
selectize_Stocks(components)

sliderinput_dates()

plot_stocks_DE()

hover_info_DE()

plot_stocks_US()

hover_info_US()
}
\description{
UI Elements
}
