% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{sliderinput_dates}
\alias{sliderinput_dates}
\alias{plot_stocks_DE}
\alias{hover_info_DE}
\alias{selectize_corona}
\alias{sliderinput_dates_corona}
\alias{plot_corona}
\alias{hover_info_corona}
\alias{selectize_Stocks_reg}
\title{UI Elements}
\usage{
sliderinput_dates()

plot_stocks_DE()

hover_info_DE()

selectize_corona()

sliderinput_dates_corona()

plot_corona()

hover_info_corona()

selectize_Stocks_reg(components1, components2)
}
\description{
UI Elements
}
